// sdk/search/search-documents/generated/models/azure/search/documents/indexes/models.d.ts
export interface SearchResourceEncryptionKey {
    /** The name of your Azure Key Vault key to be used to encrypt your data at rest. */
    keyName: string;
    /** The version of your Azure Key Vault key to be used to encrypt your data at rest. */
    keyVersion?: string;
    /** The URI of your Azure Key Vault, also referred to as DNS name, that contains the key to be used to encrypt your data at rest. An example URI might be `https://my-keyvault-name.vault.azure.net`. */
    vaultUri: string;
    /** An explicit managed identity to use for this encryption key. If not specified and the access credentials property is null, the system-assigned managed identity is used. On update to the resource, if the explicit identity is unspecified, it remains unchanged. If "none" is specified, the value of this property is cleared. */
    identity?: SearchIndexerDataIdentityUnion;
    /** An optional value indicating whether this key is a service-level key. Default is false. */
    isServiceLevelKey?: boolean;
    /** An AAD Application ID that was granted the required access permissions to the Azure Key Vault that is to be used when encrypting your data at rest. The Application ID should not be confused with the Object ID for your AAD Application. */
    applicationId?: string;
    /** The authentication key of the specified AAD application. */
    applicationSecret?: string;
}
// ...
export type SearchIndexerDataIdentityUnion = SearchIndexerDataNoneIdentity | SearchIndexerDataUserAssignedIdentity | SearchIndexerDataIdentity;
// ...
export interface IndexingSchedule {
    /** The interval of time between indexer executions. */
    interval: string;
    /** The time when an indexer should start running. */
    startTime?: Date;
}
// ...
export interface SearchIndexerDataSourceConnection {
    /** The name of the datasource. */
    name: string;
    /** The description of the datasource. */
    description?: string;
    /** The type of the datasource. */
    type: SearchIndexerDataSourceType;
    /** A specific type of the data source, in case the resource is capable of different modalities. For example, 'MongoDb' for certain 'cosmosDb' accounts. */
    readonly subType?: string;
    /** The data container for the datasource. */
    container: SearchIndexerDataContainer;
    /** An explicit managed identity to use for this datasource. If not specified and the connection string is a managed identity, the system-assigned managed identity is used. If not specified, the value remains unchanged. If "none" is specified, the value of this property is cleared. */
    identity?: SearchIndexerDataIdentityUnion;
    /** Ingestion options with various types of permission data. */
    indexerPermissionOptions?: IndexerPermissionOption[];
    /** The data change detection policy for the datasource. */
    dataChangeDetectionPolicy?: DataChangeDetectionPolicyUnion;
    /** The data deletion detection policy for the datasource. */
    dataDeletionDetectionPolicy?: DataDeletionDetectionPolicyUnion;
    /** The ETag of the data source. */
    eTag?: string;
    /** A description of an encryption key that you create in Azure Key Vault. This key is used to provide an additional level of encryption-at-rest for your datasource definition when you want full assurance that no one, not even Microsoft, can decrypt your data source definition. Once you have encrypted your data source definition, it will always remain encrypted. The search service will ignore attempts to set this property to null. You can change this property as needed if you want to rotate your encryption key; Your datasource definition will be unaffected. Encryption with customer-managed keys is not available for free search services, and is only available for paid services created on or after January 1, 2019. */
    encryptionKey?: SearchResourceEncryptionKey;
    /** The connection string for the datasource. Set to `<unchanged>` (with brackets) if you don't want the connection string updated. Set to `<redacted>` if you want to remove the connection string value from the datasource. */
    connectionString?: string;
}
// ...
export type SearchIndexerDataSourceType = string;
// ...
export interface SearchIndexerDataContainer {
    /** The name of the table or view (for Azure SQL data source) or collection (for CosmosDB data source) that will be indexed. */
    name: string;
    /** A query that is applied to this data container. The syntax and meaning of this parameter is datasource-specific. Not supported by Azure SQL datasources. */
    query?: string;
}
// ...
export type IndexerPermissionOption = string;
// ...
export type DataChangeDetectionPolicyUnion = HighWaterMarkChangeDetectionPolicy | SqlIntegratedChangeTrackingPolicy | DataChangeDetectionPolicy;
// ...
export type DataDeletionDetectionPolicyUnion = SoftDeleteColumnDeletionDetectionPolicy | NativeBlobSoftDeleteDeletionDetectionPolicy | DataDeletionDetectionPolicy;
// ...
export interface IndexerResyncBody {
    /** Re-sync options that have been pre-defined from data source. */
    options?: IndexerResyncOption[];
}
// ...
export type IndexerResyncOption = string;
/** The type of the keysOrIds. */
export interface DocumentKeysOrIds {
    /** document keys to be reset */
    documentKeys?: string[];
    /** datasource document identifiers to be reset */
    datasourceDocumentIds?: string[];
}
export declare function documentKeysOrIdsSerializer(item: DocumentKeysOrIds): any;
/** Represents an indexer. */
export interface SearchIndexer {
    /** The name of the indexer. */
    name: string;
    /** The description of the indexer. */
    description?: string;
    /** The name of the datasource from which this indexer reads data. */
    dataSourceName: string;
    /** The name of the skillset executing with this indexer. */
    skillsetName?: string;
    /** The name of the index to which this indexer writes data. */
    targetIndexName: string;
    /** The schedule for this indexer. */
    schedule?: IndexingSchedule;
    /** Parameters for indexer execution. */
    parameters?: IndexingParameters;
    /** Defines mappings between fields in the data source and corresponding target fields in the index. */
    fieldMappings?: FieldMapping[];
    /** Output field mappings are applied after enrichment and immediately before indexing. */
    outputFieldMappings?: FieldMapping[];
    /** A value indicating whether the indexer is disabled. Default is false. */
    isDisabled?: boolean;
    /** The ETag of the indexer. */
    eTag?: string;
    /** A description of an encryption key that you create in Azure Key Vault. This key is used to provide an additional level of encryption-at-rest for your indexer definition (as well as indexer execution status) when you want full assurance that no one, not even Microsoft, can decrypt them. Once you have encrypted your indexer definition, it will always remain encrypted. The search service will ignore attempts to set this property to null. You can change this property as needed if you want to rotate your encryption key; Your indexer definition (and indexer execution status) will be unaffected. Encryption with customer-managed keys is not available for free search services, and is only available for paid services created on or after January 1, 2019. */
    encryptionKey?: SearchResourceEncryptionKey;
    /** Adds caching to an enrichment pipeline to allow for incremental modification steps without having to rebuild the index every time. */
    cache?: SearchIndexerCache;
}
export declare function searchIndexerSerializer(item: SearchIndexer): any;
export declare function searchIndexerDeserializer(item: any): SearchIndexer;
/** Represents parameters for indexer execution. */
export interface IndexingParameters {
    /** The number of items that are read from the data source and indexed as a single batch in order to improve performance. The default depends on the data source type. */
    batchSize?: number;
    /** The maximum number of items that can fail indexing for indexer execution to still be considered successful. -1 means no limit. Default is 0. */
    maxFailedItems?: number;
    /** The maximum number of items in a single batch that can fail indexing for the batch to still be considered successful. -1 means no limit. Default is 0. */
    maxFailedItemsPerBatch?: number;
    /** A dictionary of indexer-specific configuration properties. Each name is the name of a specific property. Each value must be of a primitive type. */
    configuration?: IndexingParametersConfiguration;
}
// ...
export interface FieldMapping {
    /** The name of the field in the data source. */
    sourceFieldName: string;
    /** The name of the target field in the index. Same as the source field name by default. */
    targetFieldName?: string;
    /** A function to apply to each source field value before indexing. */
    mappingFunction?: FieldMappingFunction;
}
// ...
export interface SearchIndexerCache {
    /** A guid for the SearchIndexerCache. */
    id?: string;
    /** The connection string to the storage account where the cache data will be persisted. */
    storageConnectionString?: string;
    /** Specifies whether incremental reprocessing is enabled. */
    enableReprocessing?: boolean;
    /** The user-assigned managed identity used for connections to the enrichment cache.  If the connection string indicates an identity (ResourceId) and it's not specified, the system-assigned managed identity is used. On updates to the indexer, if the identity is unspecified, the value remains unchanged. If set to "none", the value of this property is cleared. */
    identity?: SearchIndexerDataIdentityUnion;
}
// ...
export interface SearchIndexerStatus {
    /** The name of the indexer. */
    readonly name: string;
    /** Overall indexer status. */
    readonly status: IndexerStatus;
    /** Snapshot of the indexer's cumulative runtime consumption for the service over the current UTC period. */
    readonly runtime: IndexerRuntime;
    /** The result of the most recent or an in-progress indexer execution. */
    readonly lastResult?: IndexerExecutionResult;
    /** History of the recent indexer executions, sorted in reverse chronological order. */
    readonly executionHistory: IndexerExecutionResult[];
    /** The execution limits for the indexer. */
    readonly limits: SearchIndexerLimits;
    /** All of the state that defines and dictates the indexer's current execution. */
    readonly currentState?: IndexerCurrentState;
}
export declare function searchIndexerStatusDeserializer(item: any): SearchIndexerStatus;
/** Represents the overall indexer status. */
export type IndexerStatus = "unknown" | "error" | "running";
/** Represents the indexer's cumulative runtime consumption in the service. */
export interface IndexerRuntime {
    /** Cumulative runtime of the indexer from the beginningTime to endingTime, in seconds. */
    usedSeconds: number;
    /** Cumulative runtime remaining for all indexers in the service from the beginningTime to endingTime, in seconds. */
    remainingSeconds?: number;
    /** Beginning UTC time of the 24-hour period considered for indexer runtime usage (inclusive). */
    beginningTime: Date;
    /** End UTC time of the 24-hour period considered for indexer runtime usage (inclusive). */
    endingTime: Date;
}
// ...
export interface SearchIndexerSkillset {
    /** The name of the skillset. */
    name: string;
    /** The description of the skillset. */
    description?: string;
    /** A list of skills in the skillset. */
    skills: SearchIndexerSkillUnion[];
    /** Details about the Azure AI service to be used when running skills. */
    cognitiveServicesAccount?: CognitiveServicesAccountUnion;
    /** Definition of additional projections to Azure blob, table, or files, of enriched data. */
    knowledgeStore?: SearchIndexerKnowledgeStore;
    /** Definition of additional projections to secondary search index(es). */
    indexProjection?: SearchIndexerIndexProjection;
    /** The ETag of the skillset. */
    eTag?: string;
    /** A description of an encryption key that you create in Azure Key Vault. This key is used to provide an additional level of encryption-at-rest for your skillset definition when you want full assurance that no one, not even Microsoft, can decrypt your skillset definition. Once you have encrypted your skillset definition, it will always remain encrypted. The search service will ignore attempts to set this property to null. You can change this property as needed if you want to rotate your encryption key; Your skillset definition will be unaffected. Encryption with customer-managed keys is not available for free search services, and is only available for paid services created on or after January 1, 2019. */
    encryptionKey?: SearchResourceEncryptionKey;
}